Move IndexRedirect & PrivateRoute fo common/routing
