+++ src/common/routing/IndexRedirect.js
