+++ src/common/routing/index.js
@@ -1,0 +1,7 @@
+import IndexRedirect from './IndexRedirect';
+import PrivateRoute from './PrivateRoute';
+
+export {
+  IndexRedirect,
+  PrivateRoute
+};
