+++ src/config.js
@@ -1,3 +1,3 @@
 export default {
-  apiUrl: 'http://localhost:5000',
+  apiUrl: 'http://localhost:8080',
 }
